+++ configure.json
@@ -1,5 +1,5 @@
 {
     "subconfigs": [
-        "src/avplayer"
+        "src/QtAVPlayer"
     ]
 }
