+++ src/QtAVPlayer/configure.json
@@ -1,5 +1,5 @@
 {
-    "module": "avplayer",
+    "module": "QtAVPlayer",
     "depends": [
         "multimedia"
     ],
