+++ src/QtAVPlayer/qavaudioframe.cpp
