+++ src/QtAVPlayer/qavaudioframe.h
@@ -8,8 +8,8 @@
 #ifndef QAVFAUDIORAME_H
 #define QAVFAUDIORAME_H
 
-#include <qavframe.h>
-#include <qavaudioformat.h>
+#include <QtAVPlayer/qavframe.h>
+#include <QtAVPlayer/qavaudioformat.h>
 
 QT_BEGIN_NAMESPACE
