+++ src/QtAVPlayer/qavdemuxer.cpp
@@ -9,7 +9,7 @@
 #include "qavvideocodec_p.h"
 #include "qavaudiocodec_p.h"
 #include "qavhwdevice_p.h"
-#include <qtavplayerglobal.h>
+#include <QtAVPlayer/qtavplayerglobal.h>
 
 #if QT_CONFIG(va_x11) && QT_CONFIG(opengl)
 #include "qavhwdevice_vaapi_x11_glx_p.h"
@@ -35,11 +35,10 @@ extern "C" {
 }
 #endif
 
-#include <QGuiApplication>
 #include <QAtomicInt>
-#include <QDebug>
-#include <QUrl>
+#include <QGuiApplication>
 #include <QDir>
+#include <QDebug>
 
 extern "C" {
 #include <libavformat/avformat.h>
@@ -116,13 +115,13 @@ static QAVVideoCodec *create_video_codec(AVStream *stream)
     Q_UNUSED(opts);
     auto name = QGuiApplication::platformName();
 
-#if QT_CONFIG(va_x11) && QT_CONFIG(opengl)
+#if defined(VA_X11) && QT_CONFIG(opengl)
     if (name == QLatin1String("xcb")) {
         device.reset(new QAVHWDevice_VAAPI_X11_GLX);
         av_dict_set(&opts, "connection_type", "x11", 0);
     }
 #endif
-#if QT_CONFIG(va_drm) && QT_CONFIG(egl)
+#if defined(VA_DRM) && QT_CONFIG(egl)
     if (name == QLatin1String("eglfs"))
         device.reset(new QAVHWDevice_VAAPI_DRM_EGL);
 #endif
