+++ src/QtAVPlayer/qavdemuxer_p.h
@@ -22,6 +22,7 @@
 #include "qavpacket_p.h"
 #include "qtavplayerglobal_p.h"
 #include <QObject>
+#include <QUrl>
 
 QT_BEGIN_NAMESPACE
