+++ src/QtAVPlayer/qavhwdevice_p.h
