+++ src/QtAVPlayer/qavhwdevice_videotoolbox.mm
