+++ src/QtAVPlayer/qavplayer-config_p.h.in
@@ -1,0 +1,16 @@
+#pragma once
+#define QT_FEATURE_ffmpeg 1
+#cmakedefine va_drm 1
+#cmakedefine va_x11 1
+
+#if defined(va_x11)
+#define QT_FEATURE_va_x11 1
+#else
+#define QT_FEATURE_va_x11 0
+#endif
+
+#if defined(va_drm)
+#define QT_FEATURE_va_drm 1
+#else
+#define QT_FEATURE_va_drm 0
+#endif
