+++ src/QtAVPlayer/qavvideobuffer_p.h
@@ -5,8 +5,8 @@
  * Free Qt Media Player based on FFmpeg.                 *
  *********************************************************/
 
-#ifndef QAVFAUDIORAME_P_H
-#define QAVFAUDIORAME_P_H
+#ifndef QAVVIDEOBUFFER_P_H
+#define QAVVIDEOBUFFER_P_H
 
 //
 //  W A R N I N G
@@ -19,18 +19,25 @@
 // We mean it.
 //
 
+#include <QtAVPlayer/qavvideoframe.h>
+#include <QVariant>
+
 QT_BEGIN_NAMESPACE
 
-struct AVFrame;
-class QAVFrame;
-class QAVCodec;
-class QAVFramePrivate
+class Q_AVPLAYER_EXPORT QAVVideoBuffer
 {
 public:
-    virtual ~QAVFramePrivate() = default;
+    QAVVideoBuffer() = default;
+    explicit QAVVideoBuffer(const QAVVideoFrame &frame) : m_frame(frame) { }
+    virtual ~QAVVideoBuffer() = default;
+    const QAVVideoFrame &frame() const { return m_frame; }
 
-    const QAVCodec *codec = nullptr;
-    AVFrame *frame = nullptr;
+    virtual QAVVideoFrame::MapData map() const = 0;
+    virtual QAVVideoFrame::HandleType handleType() const { return QAVVideoFrame::NoHandle; }
+    virtual QVariant handle() const { return {}; }
+
+protected:
+    QAVVideoFrame m_frame;
 };
 
 QT_END_NAMESPACE
