+++ src/src.pro
@@ -1,6 +1,6 @@
 TEMPLATE = subdirs
 
-include($$OUT_PWD/avplayer/qtavplayer-config.pri)
+include($$OUT_PWD/QtAVPlayer/qtQtAVPlayer-config.pri)
 QT_FOR_CONFIG += aplayer-private
 
-SUBDIRS += avplayer
+SUBDIRS += QtAVPlayer
