+++ src/srclib.pro
@@ -1,7 +1,7 @@
 TEMPLATE = subdirs
 
-include($$OUT_PWD/avplayer/qtavplayer-config.pri)
+include($$OUT_PWD/QtAVPlayer/qtQtAVPlayer-config.pri)
 QT_FOR_CONFIG += aplayer-private
 
-sub_src.file = avplayer/avplayerlib.pro
+sub_src.file = QtAVPlayer/QtAVPlayerlib.pro
 SUBDIRS += sub_src
